+++ numpy/src/main/java/org/bytedeco/numpy/presets/numpy.java
@@ -103,8 +103,8 @@ public class numpy implements InfoMapper {
                                                            "         PyUFunc_API\\[42\\]\\)").skip())
 
                .put(new Info("NPY_VISIBILITY_HIDDEN", "NPY_GCC_UNROLL_LOOPS", "NPY_GCC_OPT_3",
-                             "NPY_GCC_TARGET_AVX", "NPY_GCC_TARGET_AVX2", "NPY_GCC_TARGET_AVX512F", "NPY_RETURNS_BORROWED_REF",
-                             "NPY_INLINE", "NPY_TLS", "constchar",
+                             "NPY_GCC_TARGET_AVX", "NPY_GCC_TARGET_FMA", "NPY_GCC_TARGET_AVX2", "NPY_GCC_TARGET_AVX512F",
+                             "NPY_INLINE", "NPY_TLS", "NPY_RETURNS_BORROWED_REF", "constchar",
                              "NPY_SIGINT_ON", "NPY_SIGINT_OFF", "NPY_INPLACE",
                              "npy_fseek", "npy_ftell", "npy_lseek", "npy_off_t",
                              "PyIntpArrType_Type", "PyUIntpArrType_Type", "MyPyLong_FromInt64", "MyPyLong_AsInt64",